Rename WSL Neovim config options and update init.lua
